+++ chat_service/const/fire_chat_const.dart
@@ -1,7 +1,0 @@
-import 'package:flutter/material.dart';
-
-@immutable
-class FireChatConst {
-  static const String usersCollectionName = "users";
-  static const String roomsCollectionName = "rooms";
-}