+++ lib/models/messages/text_message.g.dart
@@ -15,9 +15,16 @@ TextMessage _$TextMessageFromJson(Map<String, dynamic> json) => TextMessage(
           ? null
           : PreviewData.fromJson(json['previewData'] as Map<String, dynamic>),
       remoteId: json['remoteId'] as String?,
+      // repliedMessage: json['repliedMessage'] == null
+      //     ? null
+      //     : Message.fromJson(
+      //         Map<String, dynamic>.from(json['repliedMessage'] as Map),
+      //       ),
       repliedMessage: json['repliedMessage'] == null
           ? null
-          : Message.fromJson(json['repliedMessage'] as Map<String, dynamic>),
+          : Message.fromJson(
+              WebJsonConverter.toMap(json['repliedMessage']),
+            ),
       roomId: json['roomId'] as String?,
       showStatus: json['showStatus'] as bool?,
       status: $enumDecodeNullable(_$StatusEnumMap, json['status']),